CA-102 - Changes for AC 3 - the Save button.
This commit updates the backend code to support sending an array of links to be updated. This also updates the frontend code to keep track of the pending changes to links in a Map, then convert that to an array when the Save button is clicked, and finally send that Array to the backend to update.