CA-135 Refactor styling
This commit fixes the background color of the AppBar in the concordance view alignment editor popup in dark mode. It also fixes the alignment of the AppBar in the popup so the buttons are lined up on the left. It also adds a transparent color to the palette for easy use throughout the app.